Complete Exercise 19-5
Refactored the index page to move the list of posts to be shown on a separate forum page.
Posts were all left to be publicly accessed, however only registered users may add new
posts. Owner foreignkey relationship added to blog posts. Further, any registered user
who is not the owner of a post is not allowed to edit that post. Finally, login_required
decorators were included to limit access to non-registered members.